crimson/os/seastore/rbm: move mkfs to RBMDevice
