+++ src/crimson/os/seastore/random_block_manager/block_rb_manager.h
@@ -22,61 +22,16 @@
 #include "include/buffer.h"
 #include "include/uuid.h"
 
+
 namespace crimson::os::seastore {
 
 constexpr rbm_abs_addr RBM_START_ADDRESS = 0;
-constexpr uint32_t RBM_SUPERBLOCK_SIZE = 4096;
 
 using RBMDevice = random_block_device::RBMDevice;
 using RBMDeviceRef = std::unique_ptr<RBMDevice>;
 
-enum {
-  // TODO: This allows the device to manage crc on a block by itself
-  RBM_NVME_END_TO_END_PROTECTION = 1,
-  RBM_BITMAP_BLOCK_CRC = 2,
-};
-
-struct rbm_metadata_header_t {
-  size_t size = 0;
-  size_t block_size = 0;
-  uint64_t start; // start location of the device
-  uint64_t end;   // end location of the device
-  uint64_t magic; // to indicate randomblock_manager
-  uuid_d uuid;
-  uint32_t start_data_area;
-  uint64_t flag; // reserved
-  uint64_t feature;
-  device_id_t device_id;
-  checksum_t crc;
-
-  DENC(rbm_metadata_header_t, v, p) {
-    DENC_START(1, 1, p);
-    denc(v.size, p);
-    denc(v.block_size, p);
-    denc(v.start, p);
-    denc(v.end, p);
-    denc(v.magic, p);
-    denc(v.uuid, p);
-    denc(v.start_data_area, p);
-    denc(v.flag, p);
-    denc(v.feature, p);
-    denc(v.device_id, p);
-
-    denc(v.crc, p);
-    DENC_FINISH(p);
-  }
-
-};
-
-std::ostream &operator<<(std::ostream &out, const rbm_metadata_header_t &header);
-
-}
-
-WRITE_CLASS_DENC_BOUNDED(
-  crimson::os::seastore::rbm_metadata_header_t
-)
-
-namespace crimson::os::seastore {
+device_config_t get_rbm_ephemeral_device_config(
+    std::size_t index, std::size_t num_devices);
 
 class BlockRBManager final : public RandomBlockManager {
 public:
@@ -88,7 +43,6 @@ public:
    * ---------------------------------------------------------------------------
    */
 
-  mkfs_ertr::future<> mkfs(mkfs_config_t) final;
   read_ertr::future<> read(paddr_t addr, bufferptr &buffer) final;
   write_ertr::future<> write(paddr_t addr, bufferptr &buf) final;
   open_ertr::future<> open() final;
@@ -110,11 +64,8 @@ public:
   abort_allocation_ertr::future<> abort_allocation(Transaction &t) final;
   write_ertr::future<> complete_allocation(Transaction &t) final;
 
-  read_ertr::future<rbm_metadata_header_t> read_rbm_header(rbm_abs_addr addr);
-  write_ertr::future<> write_rbm_header();
-
-  size_t get_size() const final { return super.size; };
-  extent_len_t get_block_size() const final { return super.block_size; }
+  size_t get_size() const final { return device->get_available_size(); };
+  extent_len_t get_block_size() const final { return device->get_block_size(); }
 
   /*
    * We will have mulitple partitions (circularjournals and randbomblockmanagers)
@@ -127,12 +78,14 @@ public:
   write_ertr::future<> write(rbm_abs_addr addr, bufferlist &bl);
 
   device_id_t get_device_id() const final {
-    return super.device_id;
+    assert(device);
+    return device->get_device_id();
   }
 
   uint64_t get_free_blocks() const final { 
     // TODO: return correct free blocks after block allocator is introduced
-    return super.size / super.block_size;
+    assert(device);
+    return get_size() / get_block_size();
   }
 
 private:
@@ -140,7 +93,6 @@ private:
    * this contains the number of bitmap blocks, free blocks and
    * rbm specific information
    */
-  rbm_metadata_header_t super;
   //FreelistManager free_manager; // TODO: block management
   RBMDevice * device;
   std::string path;
